% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockseg-class.R
\docType{methods}
\name{plot,blockSeg-method}
\alias{plot,blockSeg-method}
\title{Plot method for a blockSeg object}
\usage{
\S4method{plot}{blockSeg}(x, y, lambda = NULL, ask = TRUE,
  col = "GrayLevel", ...)
}
\arguments{
\item{x}{an object of class \code{blockSeg}.}

\item{y}{used  for S4 compatibility.}

\item{lambda}{parameter used in the LASSO.}

\item{ask}{If \code{TRUE}, to hit will be necessary to see next plot.}

\item{col}{for the colors of the representations}

\item{...}{used for S4 compatibility.}
}
\value{
a \pkg{ggplot2} object which can be plotted via the \code{print} method.
}
\description{
Produce a plot of two-dimensional segmentation of a \code{blockSeg} fit.
}
\seealso{
\code{\linkS4class{blockSeg}}.
}
