% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabsel-class.R
\docType{class}
\name{stab.blockSeg-class}
\alias{stab.blockSeg-class}
\alias{print,stab.blockSeg-method}
\alias{show,stab.blockSeg-method,}
\title{Class "stab.blockSeg"}
\description{
Class of object returned by the \code{stab.blockSeg} function.
}
\section{Slots}{
 \describe{

\item{\code{RowBreaks}: }{a vectors of length the number of rows. Each case contains the number of active variable identified
along the stability selection.}  

\item{\code{ColBreaks}: }{a vectors of length the number of columns. Each case contains the number of active variable identified
along the stability selection.}

}
}

\section{Methods}{

Specific plotting and predict methods are available and documented
(\code{\link{plot,stab.blockSeg-method}}, \code{\link{evolution,stab.blockSeg-method}}.
}

\seealso{
See also \code{\link{plot,stab.blockSeg-method}}, \code{\link{evolution,stab.blockSeg-method}}
\code{\link{print,blockSeg-method}} and \code{\link{stab.blockSeg}}.
}
\keyword{class}
